1.4.5

+++ YPlayer/src/controllers/API.as
@@ -14,6 +14,7 @@ package controllers
 	import com.pamakids.utils.NodeUtil;
 	import com.pamakids.utils.ObjectUtil;
 	import com.pamakids.utils.Singleton;
+	import com.pamakids.utils.URLUtil;
 	import com.plter.air.windows.utils.NativeCommand;
 	import com.plter.air.windows.utils.ShowCmdWindow;
 	import com.youli.nativeApplicationUpdater.NativeApplicationUpdater;
@@ -36,6 +37,7 @@ package controllers
 	import flash.utils.Dictionary;
 	import flash.utils.Timer;
 
+	import mx.formatters.DateFormatter;
 	import mx.utils.UIDUtil;
 
 	import models.InsertVO;
@@ -65,7 +67,8 @@ package controllers
 		private var nowOffset:Number=0;
 		private var config:Object;
 		private var autoUpadte:Boolean;
-		private var serial_number:String;
+		public var serial_number:String;
+		private var day:Number;
 
 		public function API()
 		{
@@ -88,7 +91,7 @@ package controllers
 				if (o.version)
 					so.data.version=o.version;
 				else
-					so.data.version='1.4.3';
+					so.data.version='1.4.5';
 			so.flush();
 			version=so.data.version;
 			so=SharedObject.getLocal('yp');
@@ -106,7 +109,7 @@ package controllers
 			{
 				var file:File=File.applicationStorageDirectory.resolvePath('log');
 				if (file.exists)
-					trace(FileManager.readFile(file.nativePath, false, true));
+					trace('log:' + FileManager.readFile(file.nativePath, false, true));
 				ServiceBase.HOST='http://localhost:18080/api';
 			}
 			else
@@ -199,7 +202,7 @@ package controllers
 						callback(0)
 					PAlert.show('更新失败，请检查网络连接');
 				}
-			}, newVersion + '.swf', null, callback, false, 'binary', function(e:Event, par:Object):void
+			}, 'update/' + newVersion + '.swf', null, callback, false, 'binary', function(e:Event, par:Object):void
 			{
 				if (autoUpadte)
 				{
@@ -226,48 +229,28 @@ package controllers
 //			}
 		}
 
-		public function reboot():void
+		private function get formatedNow():String
 		{
-//					var nativeProcessStartupInfo:NativeProcessStartupInfo=new NativeProcessStartupInfo();
-//					var file:File=new File();
+			var df:DateFormatter=new DateFormatter('YY-MM-DD HH:MM:SS');
+			return df.format(now);
+		}
 
-			var n:NativeCommand=new NativeCommand();
-			var args:Vector.<String>=new Vector.<String>;
-			args.push(File.applicationDirectory.resolvePath("乐播.exe").nativePath);
-
-			n.runCmd(args, ShowCmdWindow.HIDE, 1000);
-//					nativeProcessStartupInfo.executable=file;
-//					var process:NativeProcess=new NativeProcess();
-//					process.addEventListener(ProgressEvent.STANDARD_OUTPUT_DATA, function onOutputData(event:Object):void
-//					{
-//						var stdOut=process.standardOutput;
-//						var data=stdOut.readUTFBytes(process.standardOutput.bytesAvailable);
-//						PAlert.show(data);
-//					});
-//					process.start(nativeProcessStartupInfo);
-
-			NativeApplication.nativeApplication.exit();
-
-//			var mgr:ProductManager=new ProductManager("airappinstaller");
-//			var s:String="-launch " + NativeApplication.nativeApplication.applicationID + " " + NativeApplication.nativeApplication.publisherID;
-//			trace(s);
-//			mgr.launch(s);
-
-
-//			processArgs.push("hello");
-//			nativeProcessStartupInfo.arguments=processArgs;
-
-//			PAlert.show('重启了');
-
-
-//			var arr:Array=NativeApplication.nativeApplication.openedWindows;
-//			if (arr && arr.length)
-//			{
-//				for each (var w:NativeWindow in arr)
-//				{
-//					w.close();
-//				}
-//			}
+		public function reboot(forUpdate:Boolean=true):void
+		{
+			var logvo:LogVO;
+			if (forUpdate)
+				logvo=new LogVO(LogVO.REBOOT_AUTO_UPDATE, formatedNow, '自动重启为更新：' + newVersion)
+			else
+				logvo=new LogVO(LogVO.REBOOT_BY_COMMAND, formatedNow, '通过命令自动重启')
+			recordLog(logvo, function():void
+			{
+				appendLog('软件重启');
+				var n:NativeCommand=new NativeCommand();
+				var args:Vector.<String>=new Vector.<String>;
+				args.push(File.applicationDirectory.resolvePath("乐播.exe").nativePath);
+				n.runCmd(args, ShowCmdWindow.HIDE, 1000);
+				NativeApplication.nativeApplication.exit();
+			});
 		}
 
 
@@ -283,6 +266,7 @@ package controllers
 					nowOffset=date.getTime() - now.getTime();
 					setYPData('startup', now.getTime());
 					setYPData('refreshTime', now.getTime());
+					day=now.day;
 				}
 				trace('Now Offset:' + nowOffset);
 			});
@@ -362,7 +346,7 @@ package controllers
 			if (autoUpadte && refreshTimer.currentCount % 3600 == 0)
 			{
 				if (needReboot && !playingSong)
-					reboot();
+					reboot(rebootByCommand);
 				else
 					checkUpdate();
 			}
@@ -413,8 +397,22 @@ package controllers
 							FileManager.saveFile('bros.yp', vo.results.bros)
 						initBroadcasts();
 					}
-					if (compareMenus(menus, vo.results.menus as Array))
+					var daychanged:Boolean;
+					if (day != now.day && !initializing && !playingSong)
 					{
+						trace('Day Changed');
+						daychanged=true;
+						day=now.day;
+					}
+					if (compareMenus(menus, vo.results.menus as Array) || daychanged)
+					{
+						var playingValid:Boolean=checkPlayingValid();
+						if (daychanged || !playingValid)
+						{
+							menu=null;
+							songs=null;
+							songDMDic=null;
+						}
 						initMenu();
 					}
 					else if (brosChanged)
@@ -436,6 +434,11 @@ package controllers
 						mw.title=vo.results.message.title;
 						mw.open();
 					}
+					if (vo.results.reboot)
+					{
+						needReboot=true;
+						rebootByCommand=true;
+					}
 				}
 				else
 				{
@@ -443,6 +446,8 @@ package controllers
 				}
 			}, {startup: getYPData('startup'), version: version, playing: pn, serial_number: serial_number});
 		}
+
+		private var rebootByCommand:Boolean;
 
 		public var playingInfo:String;
 
@@ -468,7 +473,7 @@ package controllers
 			}, {_id: messageID, received: status == 1});
 		}
 
-		public function recordLog(o:LogVO):void
+		public function recordLog(o:LogVO, callbck:Function=null):void
 		{
 			var so:SharedObject=SharedObject.getLocal('log');
 			var logs:Array=so.data.logs;
@@ -487,6 +492,8 @@ package controllers
 					so.clear();
 				else
 					appendLog('RecordLogError:' + vo.errorResult);
+				if (callbck != null)
+					callbck();
 			}, {logs: JSON.stringify(logs), version: version, serial_number: serial_number});
 		}
 
@@ -768,6 +775,8 @@ package controllers
 
 		private function dateValidate(begin:Object, end:Object):Boolean
 		{
+			if (!begin || !end)
+				return false;
 			if (!(begin is Date))
 				begin=NodeUtil.getLocalDate(begin as String);
 			if (!(end is Date))
@@ -806,9 +815,33 @@ package controllers
 		private function checkPlayingValid():Boolean
 		{
 			var b:Boolean=true;
-			if (!menuValid(menu))
+			var clearedSize:Number=0;
+			var f:File;
+			var clearInfo:String='';
+			var o:Object;
+			var so:SharedObject=cachedSO;
+			var arr:Array=so.data.menus;
+			if (menu && !menuValid(menu))
 			{
+				arr.splice(arr.indexOf(menu._id), 1);
+				clearInfo+='清空了歌单：' + menu.name + ' ';
+				if (songs && songs.length)
+				{
+					for each (o in songs)
+					{
+						if (o.url && o.url.indexOf('http') != -1)
+						{
+							f=new File(FileManager.savedDir + URLUtil.getCachePath(o.url));
+							trace(f.name, f.size);
+							clearedSize+=f.size;
+							f.deleteFileAsync()
+						}
+					}
+				}
 				b=false;
+				menu=null;
+				songs=null;
+				songDMDic=null;
 			}
 			if (dmMenus && dmMenus.length)
 			{
@@ -817,30 +850,42 @@ package controllers
 					if (!menuValid(dm))
 					{
 						b=false;
-						break;
+						arr.splice(arr.indexOf(dm._id), 1);
+						clearInfo+='清空了DM：' + dm.name + ' ';
+						if (dm.dm_list && dm.dm_list.length)
+						{
+							for each (o in dm.dm_list)
+							{
+								if (o.url && o.url.indexOf('http') != -1)
+								{
+									f=new File(FileManager.savedDir + URLUtil.getCachePath(o.url));
+									clearedSize+=f.size;
+									f.deleteFileAsync()
+								}
+							}
+						}
 					}
 				}
 			}
-			if (FileManager.savedDir && !b)
+			if (!b)
 			{
-				var f:File=new File(FileManager.savedDir + 'cache');
-				if (f.exists)
-				{
-					var size:Number=f.size;
-					f.deleteDirectory(true);
-					var cached:SharedObject=cachedSO;
-					cached.data.menus=[];
-					cached.flush();
-					recordLog(new LogVO(LogVO.CLEAR_CACHE, Math.round(size / 1024) + '', '自动清空了过期歌单' + menu.name));
-				}
+				so.data.menus=arr;
+				so.flush();
+				trace(clearedSize / 1024, clearInfo);
+				recordLog(new LogVO(LogVO.CLEAR_CACHE, Math.round(clearedSize / 1024) + '', clearInfo));
 			}
 			return b;
 		}
 
 		private var dmMenus:Array;
 
+		private var initializing:Boolean;
+
 		private function initMenu():void
 		{
+			if (initializing)
+				return;
+			initializing=true;
 			var menus:Array=FileManager.readFile('menus.yp') as Array;
 			if (menus && menus.length)
 			{
@@ -972,6 +1017,7 @@ package controllers
 			var songDMDic:Dictionary=new Dictionary();
 //			this.dmMenu=dmMenu;
 			parseBroadcasts();
+			var ivo:InsertVO;
 			var a:Array=[];
 			if (dmMenu && dmMenu.dm_list)
 			{
@@ -983,7 +1029,7 @@ package controllers
 						if (dm.day.indexOf(day) == -1)
 							continue;
 					}
-					var ivo:InsertVO=new InsertVO();
+					ivo=new InsertVO();
 					ivo._id=dm.dm._id;
 					ivo.size=dm.size;
 					ivo.name=dm.dm.name;
@@ -1008,7 +1054,7 @@ package controllers
 				var ed:Date=DateUtil.getDateByHHMMSS('22:45:00');
 				while (bd.getTime() < ed.getTime())
 				{
-					var ivo:InsertVO=new InsertVO();
+					ivo=new InsertVO();
 					ivo.name=insertBro.name;
 					ivo.url=insertBro.url;
 					ivo.playTime=DateUtil.clone(bd);
@@ -1047,7 +1093,6 @@ package controllers
 							song._id=s._id;
 							song.url=QNService.HOST + s.url + '?p/1/avthumb/mp3/ab/' + o.quality + 'k';
 //							song.url=QNService.HOST + s.url;
-							//										song.cover = QNService.getQNThumbnail(s.cover, 200, 200);
 							song.name=s.name
 							song.duration=s.duration;
 							arr.push(song);
@@ -1123,12 +1168,21 @@ package controllers
 				if (menu && dmMenu)
 				{
 					if (hasCached(menu._id) && hasCached(dmMenu._id))
+					{
+						initializing=false;
 						return;
+					}
 				}
 				else if (menu && hasCached(menu._id))
+				{
+					initializing=false;
 					return;
+				}
 				else if (dmMenu && hasCached(dmMenu._id))
+				{
+					initializing=false;
 					return;
+				}
 			}
 
 			progress='开始初始化内容';
@@ -1147,6 +1201,7 @@ package controllers
 			}
 			pv.addEventListener('loaded', function(e:ODataEvent):void
 			{
+				initializing=false;
 				progress='';
 				if (e.data)
 				{
@@ -1282,6 +1337,7 @@ package controllers
 
 		private function noPlayList():void
 		{
+			initializing=false;
 			PAlert.show('您的播放歌单尚未准备完毕，请联系客服进行添加并保持网络连接', '初始化失败', null, function():void
 			{
 				online=true
@@ -1441,6 +1497,7 @@ package controllers
 		{
 			if (!ServiceBase.id || !QNService.token)
 				return;
+			trace('CheckLog');
 			var file:File=File.applicationStorageDirectory.resolvePath('log');
 			if (file.exists && file.isDirectory)
 			{
@@ -1452,7 +1509,7 @@ package controllers
 					QNService.instance.upload(f, function(r:Object):void
 					{
 						var re:ResultVO=r as ResultVO;
-						if (re && re.status)
+						if (re && re.status && f.exists)
 							f.deleteFile();
 					}, {key: upName});
 				}
